+++ client/src/hooks/Input/useQueryParams.ts
@@ -2,24 +2,17 @@ import { useEffect, useCallback, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
 import { useSearchParams } from 'react-router-dom';
 import { QueryClient, useQueryClient } from '@tanstack/react-query';
-import {
-  QueryKeys,
-  EModelEndpoint,
-  isAgentsEndpoint,
-  tQueryParamsSchema,
-  isAssistantsEndpoint,
-  PermissionBits,
-} from 'librechat-data-provider';
+import { QueryKeys, EModelEndpoint, PermissionBits } from 'librechat-data-provider';
 import type {
   AgentListResponse,
   TEndpointsConfig,
   TStartupConfig,
   TPreset,
 } from 'librechat-data-provider';
-import type { ZodAny } from 'zod';
 import {
   clearModelForNonEphemeralAgent,
   removeUnavailableTools,
+  processValidSettings,
   getModelSpecIconURL,
   getConvoSwitchLogic,
   logger,
@@ -28,62 +21,6 @@ import { useAuthContext, useAgentsMap, useDefaultConvo, useSubmitMessage } from 
 import { useChatContext, useChatFormContext } from '~/Providers';
 import { useGetAgentByIdQuery } from '~/data-provider';
 import store from '~/store';
-
-/**
- * Parses query parameter values, converting strings to their appropriate types.
- * Handles boolean strings, numbers, and preserves regular strings.
- */
-const parseQueryValue = (value: string) => {
-  if (value === 'true') {
-    return true;
-  }
-  if (value === 'false') {
-    return false;
-  }
-  if (!isNaN(Number(value))) {
-    return Number(value);
-  }
-  return value;
-};
-
-/**
- * Processes and validates URL query parameters using schema definitions.
- * Extracts valid settings based on tQueryParamsSchema and handles special endpoint cases
- * for assistants and agents.
- */
-export const processValidSettings = (queryParams: Record<string, string>) => {
-  const validSettings = {} as TPreset;
-
-  Object.entries(queryParams).forEach(([key, value]) => {
-    try {
-      const schema = tQueryParamsSchema.shape[key] as ZodAny | undefined;
-      if (schema) {
-        const parsedValue = parseQueryValue(value);
-        const validValue = schema.parse(parsedValue);
-        validSettings[key] = validValue;
-      }
-    } catch (error) {
-      console.warn(`Invalid value for setting ${key}:`, error);
-    }
-  });
-
-  if (
-    validSettings.assistant_id != null &&
-    validSettings.assistant_id &&
-    !isAssistantsEndpoint(validSettings.endpoint)
-  ) {
-    validSettings.endpoint = EModelEndpoint.assistants;
-  }
-  if (
-    validSettings.agent_id != null &&
-    validSettings.agent_id &&
-    !isAgentsEndpoint(validSettings.endpoint)
-  ) {
-    validSettings.endpoint = EModelEndpoint.agents;
-  }
-
-  return validSettings;
-};
 
 const injectAgentIntoAgentsMap = (queryClient: QueryClient, agent: any) => {
   const editCacheKey = [QueryKeys.agents, { requiredPermission: PermissionBits.EDIT }];
@@ -244,13 +181,12 @@ export default function useQueryParams({
     ],
   );
 
-  /**
-   * Checks if all settings from URL parameters have been successfully applied to the conversation.
-   * Compares values from validSettings against the current conversation state, handling special properties.
-   * Returns true only when all relevant settings match the target values.
-   */
+  const conversationRef = useRef(conversation);
+  conversationRef.current = conversation;
+
   const areSettingsApplied = useCallback(() => {
-    if (!validSettingsRef.current || !conversation) {
+    const convo = conversationRef.current;
+    if (!validSettingsRef.current || !convo) {
       return false;
     }
 
@@ -259,13 +195,13 @@ export default function useQueryParams({
         continue;
       }
 
-      if (conversation[key] !== value) {
+      if (convo[key] !== value) {
         return false;
       }
     }
 
     return true;
-  }, [conversation]);
+  }, []);
 
   /**
    * Processes message submission exactly once, preventing duplicate submissions.
@@ -285,14 +221,12 @@ export default function useQueryParams({
     methods.handleSubmit((data) => {
       if (data.text?.trim()) {
         submitMessage(data);
-
-        const newUrl = window.location.pathname;
-        window.history.replaceState({}, '', newUrl);
-
-        console.log('Message submitted with conversation state:', conversation);
+        logger.log('conversation', 'Message submitted from query params');
       }
     })();
-  }, [methods, submitMessage, conversation]);
+
+    setSearchParams(new URLSearchParams(), { replace: true });
+  }, [methods, submitMessage, setSearchParams]);
 
   useEffect(() => {
     const processQueryParams = () => {
@@ -330,12 +264,9 @@ export default function useQueryParams({
       if (!startupConfig) {
         return;
       }
-      const endpointsConfig = queryClient.getQueryData<TEndpointsConfig>([QueryKeys.endpoints]);
-      if (!endpointsConfig) {
-        return;
-      }
 
       const { decodedPrompt, validSettings, shouldAutoSubmit } = processQueryParams();
+      const hasSettings = Object.keys(validSettings).length > 0;
 
       if (!shouldAutoSubmit) {
         submissionHandledRef.current = true;
@@ -343,37 +274,36 @@ export default function useQueryParams({
 
       /** Mark processing as complete and clean up as needed */
       const success = () => {
-        const paramString = searchParams.toString();
         processedRef.current = true;
-        console.log('Parameters processed successfully', paramString);
+        logger.log('conversation', 'Query parameters processed successfully');
         clearInterval(intervalId);
 
+        // Defer URL cleanup until after submission completes (processSubmission handles it)
         if (!pendingSubmitRef.current) {
           setSearchParams(new URLSearchParams(), { replace: true });
         }
       };
 
-      // Store settings for later comparison
-      if (Object.keys(validSettings).length > 0) {
+      if (hasSettings) {
         validSettingsRef.current = validSettings;
       }
 
-      // Save the prompt text for later use if needed
       if (decodedPrompt) {
         promptTextRef.current = decodedPrompt;
       }
 
       // Handle auto-submission
       if (shouldAutoSubmit && decodedPrompt) {
-        if (Object.keys(validSettings).length > 0) {
+        if (hasSettings) {
           // Settings are changing, defer submission
           pendingSubmitRef.current = true;
 
           // Set a timeout to handle the case where settings might never fully apply
           settingsTimeoutRef.current = setTimeout(() => {
             if (!submissionHandledRef.current && pendingSubmitRef.current) {
-              console.warn(
-                'Settings application timeout reached, proceeding with submission anyway',
+              logger.log(
+                'conversation',
+                'Settings application timeout, proceeding with submission',
               );
               processSubmission();
             }
@@ -397,7 +327,7 @@ export default function useQueryParams({
         submissionHandledRef.current = true;
       }
 
-      if (Object.keys(validSettings).length > 0 && !areSettingsApplied()) {
+      if (hasSettings && !areSettingsApplied()) {
         newQueryConvo(validSettings);
       }
 
@@ -420,7 +350,6 @@ export default function useQueryParams({
     setSearchParams,
     queryClient,
     processSubmission,
-    areSettingsApplied,
   ]);
 
   useEffect(() => {
@@ -435,9 +364,7 @@ export default function useQueryParams({
       return;
     }
 
-    const allSettingsApplied = areSettingsApplied();
-
-    if (allSettingsApplied) {
+    if (areSettingsApplied()) {
       settingsAppliedRef.current = true;
 
       if (pendingSubmitRef.current) {
@@ -446,7 +373,7 @@ export default function useQueryParams({
           settingsTimeoutRef.current = null;
         }
 
-        console.log('Settings fully applied, processing submission');
+        logger.log('conversation', 'Settings fully applied, processing submission');
         processSubmission();
       }
     }
